refactor: dev branch update and folder structure refactor
