--- src/app/types/types.ts
@@ -1,4 +1,0 @@
-export interface TermsType {
-  title: string;
-  value: string;
-}
