+++ src/components/common/LegalTerms/Terms&Conditions/GoverningLaws.tsx
@@ -1,19 +1,24 @@
-import { AcceptableUsePolicy } from "~/app/constants/constant";
-import { TermsType } from "~/app/types/types";
+import { GoverningLaw } from "~/components/common/LegalTerms/Terms&Conditions/constants/constant";
 
-const AcceptableUse = () => {
+interface TermsType {
+  title: string;
+  value: string;
+}
+
+
+const GoverningLaws = () => {
   return (
     <div className="self-stretch flex flex-col items-start justify-start gap-[10px]">
       <h2 className="mt-0 self-stretch relative text-inherit font-bold font-inherit sm:text-[28px] text-2xl">
-        Acceptable Use Policy
+        Governing Law and Dispute Resolution
       </h2>
       <div className="self-stretch relative inline-block text-base">
         <div className="mb-[10px]">
-          To maintain a positive and productive environment, we have established
-          the following acceptable use policy:
+          To address any disputes that may arise, the following provisions
+          apply:
         </div>
-        <ul className="m-0 pl-[21px]">
-          {AcceptableUsePolicy.map((list: TermsType) => (
+        <ul className=" pl-[21px]">
+          {GoverningLaw.map((list: TermsType) => (
             <li className="list-disc mb-[10px]" key={list.title}>
               <span>{list.title}</span> <span>{list.value}</span>
             </li>
@@ -24,4 +29,4 @@ const AcceptableUse = () => {
   );
 };
 
-export default AcceptableUse;
+export default GoverningLaws;
