+++ src/components/common/LegalTerms/Terms&Conditions/constants/constant.ts
